1346. Check If N and Its Double Exist.
